% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo_tree_functions.R
\name{build_phylo_tree}
\alias{build_phylo_tree}
\title{Construct phylogenetic trees from consensus gene loadings across iNMF K sweep}
\usage{
build_phylo_tree(nmf_results)
}
\description{
Construct phylogenetic trees from consensus gene loadings across iNMF K sweep
}
