% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iNMF_ksweep.R
\name{iNMF_ksweep}
\alias{iNMF_ksweep}
\title{Multiplies two numbers, adds a constant, and returns the result}
\usage{
iNMF_ksweep(
  seurat.object,
  assay_slot = "RNA",
  Type.thresh = 100,
  Sample.thresh = 10,
  Batch = TRUE,
  scale.factor = 10000,
  log.norm = TRUE,
  k.min = 2,
  k.max = 40,
  n.reps = 20,
  nn.pt = 0.3,
  max.cores = parallelly::availableCores(),
  output.reps = FALSE,
  return.scale.data = T
)
}
\arguments{
\item{seurat.object}{A seurat object containing raw RNA counts}

\item{assay_slot}{Which seurat assay_slot to perform NMF on, defaults to RNA counts. For advanced users, see documentation for discussion of performing NMF on SCTransformed count data.}

\item{Type.thresh}{Minimum number of cells per cell type in n Sample.thresh samples to be included for NMF}

\item{Sample.thresh}{Minimum number of samples a given cell type must be present across for NMF}

\item{return.scale.data}{}
}
\value{
a list containing k sweep consensus NMF results
}
\description{
Run iNMF via liger
}
