% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iNMF_ksweep.R
\name{LIGER.run}
\alias{LIGER.run}
\title{LIGER.run() is a helper function to run a single replicate of iNMF at a given
rank K.}
\usage{
LIGER.run(K, Liger, minibatch_size, h5_chunk_size, seed = NULL)
}
\arguments{
\item{K}{rank to use for dimensionality reduction}

\item{Liger}{input object containing counts after scaling, normalization, and
variable feature selection}

\item{minibatch_size, h5_chunk_size}{A pair of integers that define how to
subset Liger object based on dataset size *\code{minibatch_size} refers to
number of cells in Liger object format * \code{h5_chunk_size} refers to the HDF5
file format size For more information on the computational advantages of
minibatching that online iNMF uses, see iNMF paper and original online
learning paper.}

\item{seed}{initialization value for a given iteration of iNMF}
}
\description{
LIGER.run() is a helper function to run a single replicate of iNMF at a given
rank K.
}
