% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iNMF_ksweep.R
\name{iNMF_ksweep}
\alias{iNMF_ksweep}
\title{Run dimensionality reduction (iNMF) at multiple ranks k}
\usage{
iNMF_ksweep(
  seurat.object,
  assay_slot = "RNA",
  Type.thresh = 100,
  Sample.thresh = 10,
  Batch = TRUE,
  scale.factor = 10000,
  log.norm = TRUE,
  k.min = 2,
  k.max = 40,
  n.reps = 20,
  nn.pt = 0.3,
  max.cores = parallelly::availableCores(),
  output.reps = FALSE,
  return.scale.data = F
)
}
\arguments{
\item{seurat.object}{A seurat object containing raw RNA counts}

\item{assay_slot}{Which seurat assay_slot to perform NMF on, defaults to RNA
counts. For advanced users, see documentation for discussion of performing
NMF    on SCTransformed count data.}

\item{Type.thresh}{Minimum number of cells per cell type in n Sample.thresh
samples to be included for NMF}

\item{Sample.thresh}{Minimum number of samples a given cell type must be
present across for NMF}

\item{Batch}{Sequencing batch}

\item{scale.factor}{Scale.factor for count normalization}

\item{log.norm}{Perform log normalization on counts}

\item{k.min, k.max}{Max and min ranks k for parameter sweep}

\item{n.reps}{Number of reps for consensus nmf}

\item{nn.pt}{Resolution for nearest neighbors clustering for consensus nmf}

\item{max.cores}{N cores to use for parallelization.}

\item{output.reps}{Save individual NMF reps. Default FALSE.}

\item{return.scale.data}{Save scaled data. Default FALSE.}
}
\value{
a list containing k sweep consensus NMF results
}
\description{
Run dimensionality reduction (iNMF) at multiple ranks k
}
